LayoutTests/http/tests/local/formdata/upload-events-expected.txt

Test that upload events are fired when sending FormData via XMLHttpRequest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


loadstart event fired on XMLHttpRequestUpload: [object XMLHttpRequestProgressEvent]
onprogress event fired on XMLHttpRequestUpload: [object XMLHttpRequestProgressEvent]
load event fired on XMLHttpRequestUpload: [object XMLHttpRequestProgressEvent]
string1=foo&file1=file-for-drag-to-send.txt:text/plain:1234567890
PASS successfullyParsed is true

TEST COMPLETE
